+++ .gitignore
@@ -2,3 +2,4 @@ __tests__/runner/*
 frogbot
 frogbot.exe
 *mock*
+coverage
